System.RaspberryPi.GPIO

Description

Library for controlling the GPIO pins on a Raspberry Pi (or any system using the Broadcom 2835 SOC). It is constructed as a FFI wrapper over the BCM2835 library by Mike McCauley.

Data types

data Pin

data PinMode

A GPIO pin can be either set to input or output mode. Setting it to input and writing to it will ...???

type LogicLevel = Bool

Either high or low.

type Address = Word8

This describes the address of an I2C slave.

General functions

withGPIO :: IO a -> IO a

GPIO specific functions

setPinFunction :: Pin -> PinMode -> IO ()

readPin :: Pin -> IO LogicLevel

writePin :: Pin -> LogicLevel -> IO ()

I2C specific functions

withI2C :: IO a -> IO a

setI2cAddress :: Address -> IO ()

setI2cClockDivider :: Word16 -> IO ()

setI2cBaudRate :: Word32 -> IO ()

writeI2C :: Address -> ByteString -> IO (Either ByteString ())

readI2C :: Address -> Int -> IO (Either ByteString ByteString)

writeReadI2C :: Address -> ByteString -> Int -> IO (Either ByteString ByteString)